feat(cancelOpt): widely-covered commands

+++ src/commands/user/gacha.js
@@ -80,7 +80,11 @@ class Gacha extends Command {
             socket: {emoji: emoji(`lucky_ticket`)}
         })
 		this.addConfirmationButton(`suggestToBuy`, this.suggestToBuy)
-        this.confirmationButtons.get(`suggestToBuy`).on(`collect`, async () => {
+        this.confirmationButtons.get(`suggestToBuy`).on(`collect`, async r1 => {
+			//  Handle cancellation
+			if (this.isCancelled(r1)) return reply(this.locale.ACTION_CANCELLED, {
+				socket: {emoji: emoji(`AnnieSleep`)}
+			})
             this.suggestToBuy.delete()
             this.insufficientTicketWarning.delete()
 
@@ -99,7 +103,11 @@ class Gacha extends Command {
             })
 
             this.addConfirmationButton(`checkout`, this.checkout)
-            this.confirmationButtons.get(`checkout`).on(`collect`, async () => {
+            this.confirmationButtons.get(`checkout`).on(`collect`, async r2 => {
+                //  Handle cancellation
+                if (this.isCancelled(r2)) return reply(this.locale.ACTION_CANCELLED, {
+                    socket: {emoji: emoji(`AnnieSleep`)}
+                })
                 this.checkout.delete()
                 //  Deduct balance & deliver lucky tickets
                 await this.bot.db.updateInventory({itemId: paymentItem.item_id, value: amountToPay, operation: `-`, userId: this.user.id, guildId: this.message.guild.id})
@@ -119,7 +127,11 @@ class Gacha extends Command {
                  */
                 this.askToOpenGacha = await reply(this.locale.GACHA.ASK_TO_OPEN, {simplified: true})
                 this.addConfirmationButton(`askToOpenGacha`, this.askToOpenGacha)
-                this.confirmationButtons.get(`askToOpenGacha`).on(`collect`, async () => {
+                this.confirmationButtons.get(`askToOpenGacha`).on(`collect`, async r3 => {
+                        //  Handle cancellation
+                        if (this.isCancelled(r3)) return reply(this.locale.ACTION_CANCELLED, {
+                            socket: {emoji: emoji(`AnnieSleep`)}
+                        })
                         this.askToOpenGacha.delete()
                         this.startsRoll()
                 })
